src/aws/mves/s3-minio-delta/pyproject.toml
[project]
name = "minio-docker-delta"
version = "0.1.0"
description = "MinIO example with Docker and Delta Lake"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deltalake",
    "pandas",
    "pyarrow",
    "python-dotenv",
]
